+++ packages/command-create/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @lania-cli/command-create
 
+## 1.0.1
+
+### Patch Changes
+
+- 优化了一些细节方面的问题
+- Updated dependencies
+  - @lania-cli/templates@1.0.1
+  - @lania-cli/common@1.0.1
+  - @lania-cli/compilers@1.0.1
+  - @lania-cli/linters@1.0.1
+
 ## 1.0.0
 
 ### Major Changes