+++ packages/command-dev/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @lania-cli/command-dev
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @lania-cli/common@1.0.1
+  - @lania-cli/compilers@1.0.1
+
 ## 1.0.0
 
 ### Major Changes