+++ packages/command-release/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lania-cli/command-release
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @lania-cli/common@1.0.1
+
 ## 1.0.0
 
 ### Major Changes