+++ packages/command-sync/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @lania-cli/command-sync
 
+## 1.0.1
+
+### Patch Changes
+
+- 优化了一些细节方面的问题
+- Updated dependencies
+  - @lania-cli/common@1.0.1
+
 ## 1.0.0
 
 ### Major Changes