+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lania-cli/common
 
+## 1.0.1
+
+### Patch Changes
+
+- 优化了一些细节方面的问题
+
 ## 1.0.0
 
 ### Major Changes