+++ packages/core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @lania/cli
 
+## 1.0.1
+
+### Patch Changes
+
+- 优化了一些细节方面的问题
+- Updated dependencies
+  - @lania-cli/command-create@1.0.1
+  - @lania-cli/command-sync@1.0.1
+  - @lania-cli/templates@1.0.1
+  - @lania-cli/common@1.0.1
+  - @lania-cli/command-add@1.0.1
+  - @lania-cli/command-build@1.0.1
+  - @lania-cli/command-dev@1.0.1
+  - @lania-cli/command-lint@1.0.1
+
 ## 1.0.0
 
 ### Major Changes